refactor app package tests
